Move pagination classes to separate namespace
