+++ tests/Pagination/KeysetPaginationTest.php
@@ -13,7 +13,7 @@ use Yiisoft\Factory\NotFoundException;
 use Yiisoft\Yii\DataView\Column\DataColumn;
 use Yiisoft\Yii\DataView\Exception\PaginatorNotSetException;
 use Yiisoft\Yii\DataView\GridView;
-use Yiisoft\Yii\DataView\KeysetPagination;
+use Yiisoft\Yii\DataView\Pagination\KeysetPagination;
 use Yiisoft\Yii\DataView\Tests\Support\Assert;
 use Yiisoft\Yii\DataView\Tests\Support\TestTrait;
